# web/static/openapi.yaml
openapi: 3.0.3
info:
  title: freehire public API
  version: 1.1.0
  description: |
    Public, read-only freehire API: search IT jobs, read the live filter
    vocabulary, inspect job and company details, and find similar jobs. Every
    operation here is unauthenticated. The authenticated tracking surface
    (saved jobs, application stages, CV tooling) is deliberately excluded — this
    schema is the integration contract for search, and is also what the freehire
    custom GPT imports as an Action.

    ## Filter grammar

    Facet parameters are **repeatable**, and the same grammar applies to all of
    them:

    * `?skills=go&skills=kubernetes` — values of one facet are **ORed**: match
      either.
    * `?skills=go&skills=kubernetes&skills_mode=and` — `<facet>_mode=and` flips
      that facet to **AND**: match both. Not available for geography (see below).
    * `?source_exclude=adzuna` — `<facet>_exclude` removes matches. **Every**
      string facet in this schema supports its `_exclude` twin, whether or not it
      is listed below; only the most commonly used ones are declared explicitly.
    * Different facets **AND** together: `?work_mode=remote&seniority=senior` is
      remote *and* senior.

    ## Geography is one OR-group

    `regions`, `countries` and `cities` name a single concept — *where* — so
    their values join into **one OR-group** instead of intersecting.
    `?countries=gb&cities=London` therefore returns *United Kingdom **or**
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security: []
...
paths:
...
  /companies:
    get:
      operationId: searchCompanies
      tags:
        - Companies
      summary: Search companies
      description: |
        Companies that currently have at least one open role, most active first.
        Facet parameters are repeatable (OR within a facet, AND across facets)
        and compose with `q`.

        `meta.total` is exact whenever any filter is present. On a completely
        unfiltered request it is a planner **estimate**, because counting the
        whole catalogue exactly is too expensive to do per request.
      x-openai-isConsequential: false
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Case-insensitive company name match.
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - rating
          description: |
            `rating` orders by average feedback rating. Omit for the default
            order (open-role count descending, then name).
        - name: collections
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: Curated collection slugs, e.g. `yc`, `bigtech`, `unicorn`.
        - name: regions
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: Regions the company posts roles in.
        - name: countries
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: Countries the company posts roles in (ISO 3166-1 alpha-2, lowercase).
        - name: remote_regions
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: |
            Regions the company hires REMOTELY in — derived from its open
            postings, unlike `regions`, which counts any posting.
        - name: industries
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: Curated company industry slugs, e.g. `fintech`, `developer-tools`.
        - name: domains
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: Job-derived domain slugs (the `domains` job facet, aggregated per company).
        - name: company_type
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              enum:
                - product
                - startup
                - outsource
                - outstaff
                - agency
                - inhouse
                - government
        - name: company_size
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              enum:
                - 1-10
                - 11-50
                - 51-200
                - 201-500
                - 501-1000
                - 1000+
        - name: maturity
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: Curated company maturity stage.
        - name: yc_batch
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: Y Combinator batch, e.g. `W21`. From the curated YC directory.
        - name: yc_status
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: Y Combinator company status, e.g. `Active`, `Acquired`.
        - name: yc_stage
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: Y Combinator funding stage.
        - name: yc_flags
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          description: Y Combinator directory flags, e.g. `nonprofit`, `top_company`.
      responses:
        "200":
          description: Companies with open-role counts.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CompanySummary"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
        default:
          $ref: "#/components/responses/Error"
  /companies/{slug}:
    get:
      operationId: getCompany
      tags:
        - Companies
      summary: Get company details and open jobs
      x-openai-isConsequential: false
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          description: Company slug.
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Company and its open jobs.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    type: object
                    additionalProperties: true
                    properties:
                      company:
                        $ref: "#/components/schemas/CompanyDetail"
                      jobs:
                        type: array
                        items:
                          $ref: "#/components/schemas/Job"
                      referral_available:
                        type: boolean
                        description: Whether an employee referral offer exists for this company.
        default:
          $ref: "#/components/responses/Error"
...
components:
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Page size. Values above the maximum are clamped, not rejected.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: |
        Rows to skip. On the job endpoints `offset + limit` may not exceed
        10000; deeper paging is a 400.
...
  schemas:
...
    CompanySummary:
      type: object
      additionalProperties: true
      required:
        - slug
        - name
      properties:
        slug:
          type: string
        name:
          type: string
        job_count:
          type: integer
          description: Open roles currently indexed for this company.
        tagline:
          type: string
          nullable: true
        industries:
          type: array
          items:
            type: string
        hq_country:
          type: string
          nullable: true
          description: ISO 3166-1 alpha-2, lowercase.
        collections:
          type: array
          items:
            type: string
        feedback_count:
          type: integer
        feedback_rating_avg:
          type: number
          nullable: true
    CompanyDetail:
      type: object
      additionalProperties: true
      required:
        - slug
        - name
      description: The company summary plus the profile fields served only on the detail endpoint.
      properties:
        slug:
          type: string
        name:
          type: string
        job_count:
          type: integer
        tagline:
          type: string
          nullable: true
        company_info:
          type: object
          nullable: true
          additionalProperties: true
          description: Homepage, description, and other profile fields where known.
        industries:
          type: array
          items:
            type: string
        domains:
          type: array
          items:
            type: string
        collections:
          type: array
          items:
            type: string
        regions:
          type: array
          items:
            type: string
        countries:
          type: array
          items:
            type: string
        remote_regions:
          type: array
          items:
            type: string
          description: Regions the company hires remotely in, derived from its open postings.
        company_types:
          type: array
          items:
            type: string
        company_sizes:
          type: array
          items:
            type: string
        hq_country:
          type: string
          nullable: true
        year_founded:
          type: integer
          nullable: true
        employee_count:
          type: integer
          nullable: true
        organization_type:
          type: string
          nullable: true
        maturity:
          type: string
          nullable: true
        yc_batch:
          type: array
          items:
            type: string
        yc_status:
          type: array
          items:
            type: string
        yc_stage:
          type: array
          items:
            type: string
        yc_flags:
          type: array
          items:
            type: string
        feedback_count:
          type: integer
        feedback_rating_avg:
          type: number
          nullable: true
        upvote_count:
          type: integer
        downvote_count:
          type: integer
        my_vote:
          type: integer
          nullable: true
          description: The calling user's vote. Always null on this unauthenticated schema.
...
    PaginationMeta:
      type: object
      additionalProperties: true
      properties:
        total:
          type: integer
          description: |
            Postings matching the filter. Exact on the job endpoints; on an
            unfiltered `searchCompanies` request it is a planner estimate.
        limit:
          type: integer
          description: The page size actually applied, after clamping.
        offset:
          type: integer
        ignored_params:
          type: array
          description: |
            Present ONLY when the request carried a parameter no filter read —
            a clean request omits the key entirely rather than sending `[]`.
            **Check it.** A misspelled filter does not fail; it is dropped, and
            the unfiltered result that comes back is indistinguishable from a
            genuine one.
          items:
            $ref: "#/components/schemas/IgnoredParam"
    IgnoredParam:
      type: object
      required:
        - param
      properties:
        param:
          type: string
          description: The query parameter that was received and not used.
        did_you_mean:
          type: string
          description: |
            The vocabulary's name for it, when the only mistake was grammatical
            number — most facets are plural, so `country` resolves to
            `countries`. Absent when nothing is close enough; a guess would
            mislead more than silence.
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: string